Allow ScrollView to dismiss the keyboard on tap
Tapping in a text field brings up the keyboard, but (on iPhone) there is no built-in way to dismiss it. This is a problem because the keyboard may cover up other UI elements, making them inaccessible.

This fixes the problem by adding a new property, shouldEndEditingOnTap, which defaults to false. Setting it to true changes the underlying UIScrollView’s keyboardDismissMode to .onDrag.

This change also demonstrates the new behavior in the SampleApp by adding a couple of text fields to it.