Add Flow layout, for laying out elements in a flow-style layout (#489)
https://block.atlassian.net/browse/UI-6001

Introduces a `Flow` layout type, which allows for laying out elements in
a flow style – as a row is filled with items, the layout flows to a new
row, etc. This is pretty much a direct port of the
`Verticals/Restaurant/Frameworks/RSTUIComponents/Public/Sources/Flow.swift`
layout from SPOS, which I'd helped the RST team write a while back. We
now have 4+ copies of this and similar layouts floating around the POS
repo, and requests for 2 more use cases on the way, so let's make this a
standardized element instead.

Nick has graciously offered to do the POS integration, which is here
https://github.com/squareup/ios-register/pull/103940, and allows us to
delete these duplicate implementations.