Accessibility, Traits, and UITests (#541)
This PR accomplishes 3 tasks, and can be best reviewed per commit .

First we add `.toggleButton` and `.backButton` cases to
`Accessibility.Trait` exposing traits were previously private to UIKit.

Next we make an `Accessibility` Namespace for collecting Blueprint
models of UIKit accessibility features, moving them out from under
`AccessibilityElement` and from `BlueprintUICommonControls` to
`BlueprintUI`.
Type aliases on `AccessibilityElement` for `Trait`, `CustomAction`, and
`CustomContent` are included for backwards compatibility.


Finally this PR adds a couple of XCUITests to the test suite. These both
validate that the sample app is working and that we are able to output
UI that is testable under the limitations of the XCUITest framework.