feat: Keyed layout traits (#553)
This PR adds keyed layout traits that work similarly to SwiftUI's
[LayoutValueKey](https://developer.apple.com/documentation/swiftui/layoutvaluekey).
You can define keys in the style of Environment keys, and then read the
values off subelements during layout.

To support legacy layout implementations, the `SingleTraitLayout`
protocol provides the existing API. When legacy layout is removed,
layouts that have an associated trait type only need to conform to this
protocol to keep working as-is.

Decoupling the legacy `Traits` type from the `Layout` protocol will
allow us to remove the `LegacyLayout` requirements cleanly. I'll do that
in a follow-up.