Fix an issue where resizing scroll views could result in their content offset being changed erroneously (#557)
### Bug

https://github.com/user-attachments/assets/7f93f74d-0b54-4131-ac52-8970ec623452

### Notes
🐛 If you have a partial modal (or any other modal that resizes), and you
resize the modal while the keyboard is focused and scrolled away from
the top, to be taller (eg by adding content), the scroll view within
would jump up towards the top.

Westin and I had dug in a while ago and thought it was just a UIKit bug,
but it turns out it's an order of operations thing: You need to apply
the new `contentSize` to the scroll view before you make its frame
taller, otherwise UIKit's automatic `contentOffset` adjustment kicks you
back towards the top.

To fix this, I've added a `applyBeforeLayout` addition to
`ViewDescription.Config`, which will be applied before the
`LayoutAttributes` (except on initial view allocation – we still apply
the attributes first so the view's properties are sensical).