Adding safe area edge detection for the scrollableAxes behavior (#596)
Adding a `ScrollView` API to allow clients to specify the edges that
should be inset for the safe area when using the `scrollableAxes`
content inset behavior. Before this change, content that landed outside
the safe area but inside the scroll view bounds was not scrollable when
using
[`scrollableAxes`](https://developer.apple.com/documentation/uikit/uiscrollview/contentinsetadjustmentbehavior-swift.enum/scrollableaxes).
(See linked docs for details)

The goal of this update is to address the issue of content not being
scrollable without introducing breaking API changes or large behavioral
changes to ScrollView. A few alternative approaches could be:
- turn on `alwaysBouncesVertical` and/or `alwaysBounceHorizontal`, which
will cause `scrollableAxes` to honor the safe areas along both edges of
the bouncing axes.
- update client implementations to use the `always` content inset
behavior. This would require refactoring the ScrollView layout to no
longer produce a `contentSize` equal to the ScrollView bounds when using
`fittingWidth` and `fittingHeight`.

The approach in this PR will enable the safe area behavior we need along
a precise edge, like the bottom, while limiting the risk of regressions.